+++ components/core/HelperText/HelperText.tsx
@@ -1,9 +1,35 @@
-import React from 'react'
+import warning from '@/app/_asset/icons/warning-circle.svg';
+import error from '@/app/_asset/icons/error-circle.svg';
+import success from '@/app/_asset/icons/success-circle.svg';
+import Image from 'next/image';
 
-const HelperText = () => {
-  return (
-    <div></div>
-  )
+interface HelperTextProps {
+  variant: string | null;
+  helperText?: string;
 }
 
-export default HelperText
+const circleSvgs = {
+  warning: warning,
+  error: error,
+  success: success,
+};
+
+const HelperText = ({ variant, helperText }: HelperTextProps) => {
+  return (
+    <p className={`flex font-thin text-xs text-${variant}-60`}>
+      {variant === 'warning' || variant === 'error' || variant === 'success' ? (
+        <Image
+          priority
+          src={circleSvgs[variant]}
+          alt={variant}
+          width={16}
+          height={16}
+          className={`mr-[6px]`}
+        />
+      ) : null}
+      {helperText && helperText}
+    </p>
+  );
+};
+
+export default HelperText;
